refactor(user): user-related folder organization
Organization of user-related folders within the infra folder
